Patched faulty 'already signed in' detection
Previously, if you were signed in, even with the wrong credentials, the website allowed you to visit the messaging page, and the messaging page showed an empty page even if it had some new messages. It showed an empty page because the password needs to be correct to fetch the messages from the backend.